Move queries to own file
